+++ src/Console/Output.php
@@ -57,14 +57,14 @@ class Output
      * Return the terminal width. If not an interactive shell, return default 80;
      *
      * @return int
+     *
+     * @codeCoverageIgnore
      */
     public function getTerminalWidth()
     {
         // If running on windows or not an interactive shell, just pretend it's 80
         if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN' || !$this->isInteractiveShell()) {
-            // @codeCoverageIgnoreStart
             return 80;
-            // @codeCoverageIgnoreEnd
         }
         return (int)shell_exec("tput cols");
     }
@@ -73,14 +73,14 @@ class Output
      * Return the terminal width. If not an interactive shell, return default 25;
      *
      * @return int
+     *
+     * @codeCoverageIgnore
      */
     public function getTerminalHeight()
     {
         // If running on windows or not an interactive shell, just pretend it's 25
         if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN' || !$this->isInteractiveShell()) {
-            // @codeCoverageIgnoreStart
             return 25;
-            // @codeCoverageIgnoreEnd
         }
         return (int)shell_exec("tput lines");
     }
@@ -89,6 +89,8 @@ class Output
      * Return whether we're currently in an interactive shell or not.
      *
      * @return bool
+     *
+     * @codeCoverageIgnore
      */
     public function isInteractiveShell()
     {